[FEAT] 리뷰 기능 리팩토링, Swagger 적용 (#136)
* feat: 리뷰 수정, 삭제 시 접근 권한 검사 추가

* feat: swagger 적용

* fix: 리뷰 권한 로직 수정

* refactor: 리팩토링

* refactor: 리뷰 접근 로직 리팩토링

fix: hasReviewAccessPermission 사용 수정

* refactor: 상품 리뷰 작성 기능 리팩토링

* fix: 테스트 코드 컴파일 에러 수정

+++ src/main/java/com/kt/service/ReviewServiceImpl.java
@@ -12,9 +12,14 @@ import com.kt.constant.OrderStatus;
 import com.kt.constant.message.ErrorCode;
 import com.kt.constant.searchtype.ProductSearchType;
 import com.kt.domain.dto.response.ReviewResponse;
+import com.kt.domain.entity.AbstractAccountEntity;
+import com.kt.domain.entity.OrderEntity;
 import com.kt.domain.entity.OrderProductEntity;
 import com.kt.domain.entity.ReviewEntity;
+import com.kt.domain.entity.UserEntity;
 import com.kt.exception.CustomException;
+import com.kt.repository.OrderRepository;
+import com.kt.repository.account.AccountRepository;
 import com.kt.repository.orderproduct.OrderProductRepository;
 import com.kt.repository.review.ReviewRepository;
 
@@ -27,9 +32,21 @@ public class ReviewServiceImpl implements ReviewService {
 
 	private final ReviewRepository reviewRepository;
 	private final OrderProductRepository orderProductRepository;
+	private final AccountRepository accountRepository;
+	private final OrderRepository orderRepository;
 
 	@Override
-	public void create(UUID orderProductId, String content) {
+	public void create(
+		String email,
+		UUID orderProductId,
+		String content
+	) {
+		if (isOrderProductOwnedByUser(email, orderProductId))
+			throw new CustomException(ErrorCode.REVIEW_ACCESS_NOT_ALLOWED);
+
+		if (reviewRepository.findByOrderProductId(orderProductId).isPresent())
+			throw new CustomException(ErrorCode.REVIEW_ALREADY_EXISTS);
+
 		OrderProductEntity orderProduct = orderProductRepository.findByIdOrThrow(orderProductId);
 		if (orderProduct.getOrder().getStatus() != OrderStatus.PURCHASE_CONFIRMED)
 			throw new CustomException(ErrorCode.ORDER_NOT_CONFIRMED);
@@ -40,14 +57,25 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	@Override
-	public void update(UUID reviewId, String content) {
+	public void update(
+		String email,
+		UUID reviewId,
+		String content
+	) {
 		ReviewEntity review = reviewRepository.findByIdOrThrow(reviewId);
+		if (!hasReviewAccessPermission(email, review))
+			throw new CustomException(ErrorCode.REVIEW_ACCESS_NOT_ALLOWED);
 		review.update(content);
 	}
 
 	@Override
-	public void delete(UUID reviewId) {
+	public void delete(
+		String email,
+		UUID reviewId
+	) {
 		ReviewEntity review = reviewRepository.findByIdOrThrow(reviewId);
+		if (!hasReviewAccessPermission(email, review))
+			throw new CustomException(ErrorCode.REVIEW_ACCESS_NOT_ALLOWED);
 		review.delete();
 	}
 
@@ -61,30 +89,45 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	@Override
-	public List<ReviewResponse.Search> getReviewByProductId(UUID productId) {
-		List<OrderProductEntity> orderProducts = orderProductRepository.findAllByProductId(productId);
-
-		return orderProducts.stream().map(orderProduct -> {
-			ReviewEntity reviewEntity = reviewRepository.findByOrderProductIdOrThrow(orderProduct.getId());
-			return new ReviewResponse.Search(
-				reviewEntity.getId(),
-				reviewEntity.getContent()
-			);
-		}).toList();
-	}
-
-	@Override
-	public Page<ReviewResponse.Search> getReviewByProductId(UUID productId, Pageable pageable){
+	public Page<ReviewResponse.Search> getReviewByProductId(
+		UUID productId,
+		Pageable pageable
+	) {
 		return reviewRepository.searchReviewsByProductId(pageable, productId);
 	}
 
 	@Override
-	public Page<ReviewResponse.Search> getReviewsByAdmin(Pageable pageable, String keyword, ProductSearchType type){
-		return reviewRepository.searchReviews(pageable,keyword,type);
+	public Page<ReviewResponse.Search> getReviewsByAdmin(
+		Pageable pageable,
+		String keyword,
+		ProductSearchType type
+	) {
+		return reviewRepository.searchReviews(pageable, keyword, type);
 	}
 
 	@Override
 	public Page<ReviewResponse.Search> getReviewsByUserId(Pageable pageable, UUID userId) {
 		return reviewRepository.searchReviewsByUserId(pageable, userId);
+	}
+
+	private boolean isOrderProductOwnedByUser(String email, UUID orderProductId) {
+		AbstractAccountEntity user = accountRepository.findByEmailOrThrow(email);
+		List<OrderEntity> orders = orderRepository.findAllByOrderBy_Id(user.getId());
+
+		return orders
+			.stream()
+			.flatMap(order -> order.getOrderProducts().stream())
+			.anyMatch(orderProduct -> orderProduct.getId().equals(orderProductId));
+	}
+
+	private boolean hasReviewAccessPermission(String email, ReviewEntity review) {
+		AbstractAccountEntity reviewEditor = accountRepository.findByEmailOrThrow(email);
+
+		UserEntity reviewOwner = review
+			.getOrderProduct()
+			.getOrder()
+			.getOrderBy();
+
+		return reviewEditor.getEmail().equals(reviewOwner.getEmail());
 	}
 }
